requirements.txt
streamlit
yt-dlp
dropbox
python-dotenv
